Refactor: move buildDayData business logic to domain layer
Move business logic from ContributionGridStateHelpers to domain use cases:
- Create DayBuildContext model and BuildDayDataUseCase
- Expand BuildActivityDataUseCase with buildWeekData method
- Simplify presentation layer to only delegate to domain
