+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/habits/domain/usecase/BuildActivityDataUseCase.kt
@@ -1,17 +1,36 @@
 package space.be1ski.vibits.shared.feature.habits.domain.usecase
 
+import kotlinx.datetime.DatePeriod
 import kotlinx.datetime.LocalDate
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.plus
+import space.be1ski.vibits.shared.core.ui.ActivityMode
+import space.be1ski.vibits.shared.core.ui.ActivityRange
+import space.be1ski.vibits.shared.feature.habits.domain.model.ActivityWeek
 import space.be1ski.vibits.shared.feature.habits.domain.model.ActivityWeekData
 import space.be1ski.vibits.shared.feature.habits.domain.model.ContributionDay
+import space.be1ski.vibits.shared.feature.habits.domain.model.DailyMemoInfo
+import space.be1ski.vibits.shared.feature.habits.domain.model.DayBuildContext
 import space.be1ski.vibits.shared.feature.habits.domain.model.HabitConfig
 import space.be1ski.vibits.shared.feature.habits.domain.model.HabitStatus
+import space.be1ski.vibits.shared.feature.habits.domain.model.HabitsConfigEntry
+import space.be1ski.vibits.shared.feature.habits.domain.model.RangeBounds
+import space.be1ski.vibits.shared.feature.habits.domain.model.rangeBounds
+import space.be1ski.vibits.shared.feature.memos.domain.model.Memo
 
 private const val LAST_SEVEN_DAYS = 7
+private const val DAYS_IN_WEEK = 7
 
 /**
  * Use case for building and filtering activity data.
  */
-class BuildActivityDataUseCase {
+class BuildActivityDataUseCase(
+  private val buildDayDataUseCase: BuildDayDataUseCase = BuildDayDataUseCase(),
+  private val dateCalculationsUseCase: DateCalculationsUseCase = DateCalculationsUseCase(),
+  private val extractDailyMemosUseCase: ExtractDailyMemosUseCase = ExtractDailyMemosUseCase(),
+  private val countDailyPostsUseCase: CountDailyPostsUseCase = CountDailyPostsUseCase(),
+  private val extractHabitsConfigUseCase: ExtractHabitsConfigUseCase = ExtractHabitsConfigUseCase(),
+) {
   /**
    * Returns the last 7 in-range days from activity data.
    */
@@ -71,4 +90,70 @@ class BuildActivityDataUseCase {
     val maxWeekly = weeks.maxOfOrNull { it.weeklyCount } ?: 0
     return ActivityWeekData(weeks = weeks, maxDaily = maxDaily, maxWeekly = maxWeekly)
   }
+
+  /**
+   * Builds ActivityWeekData for a given range.
+   * Uses pre-extracted configTimeline and dailyMemos to avoid redundant work on range change.
+   */
+  @Suppress("LongParameterList")
+  fun buildWeekData(
+    configTimeline: List<HabitsConfigEntry>,
+    dailyMemos: Map<LocalDate, DailyMemoInfo>,
+    timeZone: TimeZone,
+    memos: List<Memo>,
+    range: ActivityRange,
+    mode: ActivityMode,
+    today: LocalDate,
+  ): ActivityWeekData {
+    val bounds = rangeBounds(range)
+    val effectiveConfigTimeline = if (mode == ActivityMode.HABITS) configTimeline else emptyList()
+    val counts = if (mode == ActivityMode.POSTS) countDailyPostsUseCase(memos, timeZone, bounds) else emptyMap()
+
+    val start = dateCalculationsUseCase.startOfWeek(bounds.start)
+    val weeks = mutableListOf<ActivityWeek>()
+    var cursor = start
+    while (cursor <= bounds.end) {
+      val days =
+        (0 until DAYS_IN_WEEK).map { offset ->
+          buildDayDataUseCase(
+            DayBuildContext(
+              date = cursor.plus(DatePeriod(days = offset)),
+              bounds = bounds,
+              mode = mode,
+              configTimeline = effectiveConfigTimeline,
+              dailyMemos = dailyMemos,
+              counts = counts,
+              today = today,
+            ),
+          )
+        }
+      weeks.add(
+        ActivityWeek(
+          startDate = cursor,
+          days = days,
+          weeklyCount = days.sumOf { it.count },
+        ),
+      )
+      cursor = cursor.plus(DatePeriod(days = DAYS_IN_WEEK))
+    }
+    val maxDaily = weeks.maxOfOrNull { week -> week.days.maxOfOrNull { it.count } ?: 0 } ?: 0
+    val maxWeekly = weeks.maxOfOrNull { it.weeklyCount } ?: 0
+    return ActivityWeekData(weeks = weeks, maxDaily = maxDaily, maxWeekly = maxWeekly)
+  }
+
+  /**
+   * Extracts habits config timeline from memos.
+   */
+  fun extractConfigTimeline(
+    memos: List<Memo>,
+    timeZone: TimeZone,
+  ): List<HabitsConfigEntry> = extractHabitsConfigUseCase(memos, timeZone)
+
+  /**
+   * Extracts daily memos map from memos.
+   */
+  fun extractDailyMemos(
+    memos: List<Memo>,
+    timeZone: TimeZone,
+  ): Map<LocalDate, DailyMemoInfo> = extractDailyMemosUseCase(memos, timeZone)
 }
